[AND-265] Move NewApp and NewAppVersion promotionals to promo folder
