+++ app-games/src/main/java/com/aptoide/android/aptoidegames/feature_promotional/NewAppPromotionalView.kt
@@ -1,5 +1,6 @@
-package com.aptoide.android.aptoidegames.home
+package com.aptoide.android.aptoidegames.feature_promotional
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -16,6 +17,17 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.clearAndSetSemantics
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import cm.aptoide.pt.download_view.presentation.DownloadUiState
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Downloading
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Error
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Install
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Installed
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Installing
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Migrate
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Outdated
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.ReadyToInstall
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Uninstalling
+import cm.aptoide.pt.download_view.presentation.DownloadUiState.Waiting
 import cm.aptoide.pt.download_view.presentation.rememberDownloadState
 import cm.aptoide.pt.feature_apps.data.App
 import cm.aptoide.pt.feature_apps.presentation.AppUiState
@@ -23,6 +35,9 @@ import cm.aptoide.pt.feature_apps.presentation.rememberApp
 import cm.aptoide.pt.feature_home.domain.Bundle
 import com.aptoide.android.aptoidegames.R
 import com.aptoide.android.aptoidegames.appview.buildAppViewRoute
+import com.aptoide.android.aptoidegames.drawables.icons.getAppcoinsClearLogo
+import com.aptoide.android.aptoidegames.home.AptoidePromotionalFeatureGraphicImage
+import com.aptoide.android.aptoidegames.home.LoadingView
 import com.aptoide.android.aptoidegames.installer.presentation.AppIconWProgress
 import com.aptoide.android.aptoidegames.installer.presentation.InstallViewShort
 import com.aptoide.android.aptoidegames.installer.presentation.ProgressText
@@ -30,12 +45,20 @@ import com.aptoide.android.aptoidegames.theme.AGTypography
 import com.aptoide.android.aptoidegames.theme.Palette
 
 @Composable
-fun NewAppVersionPromotionalView(bundle: Bundle, navigate: (String) -> Unit) {
+fun NewAppPromotionalView(
+  bundle: Bundle,
+  navigate: (String) -> Unit
+) {
   bundle.view?.let {
     val (uiState, _) = rememberApp(it)
 
     when (uiState) {
-      is AppUiState.Idle -> NewAppVersionPromotional(uiState.app, bundle.title, navigate)
+      is AppUiState.Idle -> NewAppPromotional(
+        app = uiState.app,
+        label = bundle.title,
+        navigate = navigate
+      )
+
       is AppUiState.Loading -> LoadingView()
       is AppUiState.NoConnection,
       is AppUiState.Error -> Unit
@@ -44,7 +67,7 @@ fun NewAppVersionPromotionalView(bundle: Bundle, navigate: (String) -> Unit) {
 }
 
 @Composable
-private fun NewAppVersionPromotional(app: App, label: String, navigate: (String) -> Unit) {
+private fun NewAppPromotional(app: App, label: String, navigate: (String) -> Unit) {
   val downloadState = rememberDownloadState(app)
   val isDownloading = remember(downloadState) {
     isDownloading(downloadState)
@@ -55,7 +78,11 @@ private fun NewAppVersionPromotional(app: App, label: String, navigate: (String)
       .padding(all = 16.dp)
       .clickable(onClick = { navigate(buildAppViewRoute(app)) })
   ) {
-    AptoidePromotionalFeatureGraphicImage(app.featureGraphic, label)
+    AptoidePromotionalFeatureGraphicImage(
+      featureGraphic = app.featureGraphic,
+      label = label,
+      hasAppCoins = app.isAppCoins
+    )
     Row(
       modifier = Modifier.padding(top = 8.dp),
       verticalAlignment = Alignment.CenterVertically
@@ -90,6 +117,7 @@ private fun NewAppVersionPromotional(app: App, label: String, navigate: (String)
         } else {
           AppInfo(app)
         }
+
       }
       InstallViewShort(app = app)
     }
@@ -98,31 +126,47 @@ private fun NewAppVersionPromotional(app: App, label: String, navigate: (String)
 
 @Composable
 private fun AppInfo(app: App) {
-  Row(
-    modifier = Modifier
-      .wrapContentHeight(
-        unbounded = true,
-        align = Alignment.Top
-      ),
-    verticalAlignment = Alignment.CenterVertically
-  ) {
-    Text(
+  if (app.isAppCoins) {
+    Row(
       modifier = Modifier
-        .padding(end = 8.dp),
-      text = stringResource(R.string.promotional_new_version),
-      maxLines = 1,
-      color = Palette.Primary,
-      style = AGTypography.InputsXS,
-      overflow = TextOverflow.Ellipsis,
-    )
-    Text(
-      modifier = Modifier
-        .padding(end = 8.dp),
-      text = app.versionName,
-      color = Palette.GreyLight,
-      maxLines = 1,
-      style = AGTypography.InputsXS,
-      overflow = TextOverflow.Ellipsis,
-    )
+        .wrapContentHeight(
+          unbounded = true,
+          align = Alignment.Top
+        ),
+      verticalAlignment = Alignment.CenterVertically
+    ) {
+      Image(
+        imageVector = getAppcoinsClearLogo(Palette.AppCoinsPink),
+        contentDescription = null,
+        modifier = Modifier
+          .padding(end = 8.dp)
+          .size(16.dp),
+      )
+      Text(
+        modifier = Modifier
+          .padding(end = 8.dp),
+        text = stringResource(R.string.promotional_now_with_appcoins),
+        maxLines = 1,
+        style = AGTypography.InputsXSRegular,
+        overflow = TextOverflow.Ellipsis,
+      )
+    }
+  }
+}
+
+fun isDownloading(downloadState: DownloadUiState?): Boolean {
+  return when (downloadState) {
+    is Downloading,
+    is Error,
+    is Installing,
+    is ReadyToInstall,
+    is Waiting,
+    is Uninstalling -> true
+
+    is Install,
+    is Outdated,
+    is Migrate,
+    is Installed,
+    null -> false
   }
 }
